feat: combine SG selector and detail into single view
Replace the separate security group list + detail views with a
unified 5-pane dashboard: SG selector at top, Info | Rules in
the middle, Servers | Ports at the bottom. Selecting a different
SG instantly updates all detail panes.

Context-sensitive action bar: SG CRUD on selector pane, rule
CRUD+edit on rules pane, server navigation on servers pane.
Esc from server detail returns to SG view via breadcrumb.

Removes the sgdetail package entirely — all logic merged into
secgroupview.
