+++ src/internal/app/routing.go
@@ -40,9 +40,6 @@ func (m Model) updateActiveView(msg tea.Msg) (Model, tea.Cmd) {
 	case viewSecGroupView:
 		m.secGroupView, cmd = m.secGroupView.Update(msg)
 		m.statusBar.Hint = m.secGroupView.Hints()
-	case viewSecGroupDetail:
-		m.sgDetail, cmd = m.sgDetail.Update(msg)
-		m.statusBar.Hint = m.sgDetail.Hints()
 	case viewVolumeCreate:
 		m.volumeCreate, cmd = m.volumeCreate.Update(msg)
 		m.statusBar.Hint = m.volumeCreate.Hints()
@@ -140,9 +137,6 @@ func (m Model) updateAllViews(msg tea.Msg) (Model, tea.Cmd) {
 	case viewImageDetail:
 		m.imageDetail, cmd = m.imageDetail.Update(msg)
 		cmds = append(cmds, cmd)
-	case viewSecGroupDetail:
-		m.sgDetail, cmd = m.sgDetail.Update(msg)
-		cmds = append(cmds, cmd)
 	case viewConsoleLog:
 		m.consoleLog, cmd = m.consoleLog.Update(msg)
 		cmds = append(cmds, cmd)
@@ -234,7 +228,7 @@ func (m Model) handleResourceNavigation(msg shared.NavigateToResourceMsg) (Model
 func (m Model) handleViewChange(msg shared.ViewChangeMsg) (Model, tea.Cmd) {
 	switch msg.View {
 	case "serverlist":
-		// If returning from a cross-resource jump, go back to the originating detail view
+		// If returning from a cross-resource jump, go back to the originating view
 		if m.returnToView == viewServerDetail && m.serverDetail.ServerID() != "" {
 			m.returnToView = 0
 			m.view = viewServerDetail
@@ -242,11 +236,11 @@ func (m Model) handleViewChange(msg shared.ViewChangeMsg) (Model, tea.Cmd) {
 			m.statusBar.Hint = m.serverDetail.Hints()
 			return m, nil
 		}
-		if m.returnToView == viewSecGroupDetail && m.sgDetail.SGID() != "" {
+		if m.returnToView == viewSecGroupView {
 			m.returnToView = 0
-			m.view = viewSecGroupDetail
-			m.statusBar.CurrentView = "sgdetail"
-			m.statusBar.Hint = m.sgDetail.Hints()
+			m.view = viewSecGroupView
+			m.statusBar.CurrentView = "secgroupview"
+			m.statusBar.Hint = m.secGroupView.Hints()
 			return m, nil
 		}
 		m.returnToView = 0
@@ -333,14 +327,6 @@ func (m Model) handleViewChange(msg shared.ViewChangeMsg) (Model, tea.Cmd) {
 		return m, m.serverCreate.Init()
 
 	case "secgroupview":
-		// If returning from a cross-resource jump, go back to SG detail
-		if m.returnToView == viewSecGroupDetail && m.sgDetail.SGID() != "" {
-			m.returnToView = 0
-			m.view = viewSecGroupDetail
-			m.statusBar.CurrentView = "sgdetail"
-			m.statusBar.Hint = m.sgDetail.Hints()
-			return m, nil
-		}
 		m.returnToView = 0
 		m.view = viewSecGroupView
 		m.statusBar.CurrentView = "secgroupview"
@@ -368,8 +354,6 @@ func (m Model) forceRefreshActiveView() (Model, tea.Cmd) {
 		return m, m.floatingIPList.ForceRefresh()
 	case viewSecGroupView:
 		return m, m.secGroupView.ForceRefresh()
-	case viewSecGroupDetail:
-		return m, m.sgDetail.ForceRefresh()
 	case viewKeypairList:
 		return m, m.keypairList.ForceRefresh()
 	case viewNetworkList:
